refactor(web-ui): extract capabilities, audio, attachments from api/mod.rs
Pull three small, self-contained handler groups out of the 4.4 k-line
`crates/web-ui/src/api/mod.rs` into sibling modules:

- `api/capabilities.rs` — `GET /api/capabilities` + `ServerCapabilities`
- `api/audio.rs` — `GET /api/messages/{id}/audio`, `GET /api/audio/{id}`
- `api/attachments.rs` — `POST /conversations/{id}/attachments`,
  `GET /attachments/{id}`, plus `AttachmentUploadForm`,
  `AttachmentMetaResponse`, `AttachmentServeParams`, and the
  private `resize_image` helper

Test fixtures (`StubTranscriptionProvider`, `mount_llm_reply`, `test_state`,
`event_log_state`, `app`, `body_bytes`, `body_json`, `tiny_png`,
`multipart_body`) move to a new shared `api/test_helpers.rs` so sibling
test modules can pull them via `use super::super::test_helpers::*;`.

Router calls and `openapi.rs` paths are updated to the new locations.
No behaviour changes; `cargo test -p assistant-web-ui --lib` reports
241 passed (same as before).
